Moves TestArbFunctionTable to system_tests
